+++ projects/amdsmi/src/amd_smi/amd_smi.cc
@@ -321,6 +321,10 @@ amdsmi_status_t amdsmi_shut_down() {
 }
 
 amdsmi_status_t amdsmi_status_code_to_string(amdsmi_status_t status, const char** status_string) {
+  if (status_string == nullptr) {
+    return AMDSMI_STATUS_INVAL;
+  }
+
   switch (status) {
     case AMDSMI_STATUS_SUCCESS:
       *status_string = "AMDSMI_STATUS_SUCCESS: Call succeeded.";
@@ -345,6 +349,9 @@ amdsmi_status_t amdsmi_status_code_to_string(amdsmi_status_t status, const char*
       break;
     case AMDSMI_STATUS_API_FAILED:
       *status_string = "AMDSMI_STATUS_API_FAILED: API call failed.";
+      break;
+    case AMDSMI_STATUS_TIMEOUT:
+      *status_string = "AMDSMI_STATUS_TIMEOUT: Timeout in API call.";
       break;
     case AMDSMI_STATUS_RETRY:
       *status_string = "AMDSMI_STATUS_RETRY: Retry operation.";
@@ -409,6 +416,11 @@ amdsmi_status_t amdsmi_status_code_to_string(amdsmi_status_t status, const char*
       break;
     case AMDSMI_STATUS_DRIVER_NOT_LOADED:
       *status_string = "AMDSMI_STATUS_DRIVER_NOT_LOADED: Processor driver not loaded.";
+      break;
+    case AMDSMI_STATUS_MORE_DATA:
+      *status_string =
+          "AMDSMI_STATUS_MORE_DATA: More data is available than fits in the"
+          " caller-supplied buffer.";
       break;
     case AMDSMI_STATUS_NO_DATA:
       *status_string = "AMDSMI_STATUS_NO_DATA: No data was found for a given input.";